fix: address PR review comments for CheckResultWriter refactor
- Export CheckResultWriter from analyze/utils/__init__.py
- Fix NaN check to use math.isnan() instead of obj != obj (CodeQL)
- Update check_ops.py import to go through __init__.py per CLAUDE.md
- Move top-level imports out of _make_writer helper in test file
