+++ src/winml/modelkit/analyze/utils/__init__.py
@@ -6,11 +6,13 @@
 
 from .ep_utils import infer_ihv_from_ep_name
 from .json_utils import validate_json_schema
+from .op_utils import CheckResultWriter
 from .pattern_matching import match_pattern_with_wildcards
 from .rule_loader import RuleLoader
 
 
 __all__ = [
+    "CheckResultWriter",
     "RuleLoader",
     "infer_ihv_from_ep_name",
     "match_pattern_with_wildcards",
